requirements.txt
PyYAML>=6.0.0
jinja2>=3.0.0
lunar_python>=1.3.0
python-dotenv>=1.0.0
aiosmtplib>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0 